bloghub/requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
feedparser>=6.0.0
python-dateutil>=2.8.0